pyproject.toml
[tool.poetry]
name = "ltc-insurance-api"
version = "1.0.0"
description = "FastAPI service for LTC insurance data with Snowpark integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "0.109.0"
uvicorn = {extras = ["standard"], version = "0.27.0"}
snowflake-snowpark-python = "1.11.1"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-multipart = "0.0.6"
cachetools = "5.3.2"
python-json-logger = "2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

